I made a bunch of changes.
For one, I rewrote the balls class, which I now call the Particles
class. There was some very odd stuff in this class (like improper
bounds checks + normalizing the particle velocity using the ratio of its
velocity squared to the maximum velocity, when you should have
used the ratio of the particle velocity to the maximum velocity).

I also played around with a bunch of settings to improve image sharpness.

I want to get rid of the vstack nonsense that the original author used
to keep tack of the particle histories... since we know how many
particles there are, and how many time steps each one will go through,
we can allocate static arrays to store the position histories... this
will be easier to understand/read and faster.

I also renamed a bunch of stuff.

Expect further improvements in the next commit.