I renamed a few variables + added new options.
I added an option to save the force field to file (I plan to also
write a function to read from file). I want to do this so that I can
play around with different colors for the same force field.

I also added options for DPI stuff + renamed a few variables throughout
the code.