I added a load function
This function, as the name suggests, loads a save from file.
I moved the bits of code that are common to load and run (which is all
the code starting at the bit that generates the particles) into a
new function called "simulate". I did this just to reduce code
redundancy.

I also made dedicated directories for saves and images with .gitkeep
files in them.

I also added a gitignore.