Removed edge artifacts, Rewrote Perlin, fractal.
The old code had some problems around the edges. Basically, particles
tended to cluster around the edges of the image. I wanted to fix this.
To do that, however, I first needed to figure out what was actually
causing the clustering. It was a combination of periodic BC's with
a non-periodic vector field. Imagine that on the right edge of the
grid, the vector field points to the right. On the left most edge, the
vector field points to the left. A particle travels though the right
edge, moving to the right. The periodic BCs map it to the left edge,
where the force field pushes it to the left, back across the left edge
the periodic BC's then move it to the right edge, where the force
field pushes it back to the left edge and so on. The particle basically
goes nowhere. Over time, particles cluster at this edge. Boom, there's
the artifact.

To fix this, I simply had to make the vector field periodic. To do this,
I made it so that at each layer of noise, the left most edge of the
inner grid vectors matches the left most edge (same for the top and
bottom edges). The inner points are interpolated from these and
everything becomes periodic. Woo-hoo.

To make this work, however, I had to actually figure out wtf
the perlin and fractal functions do. I did this by renaming almost
every variable and adding in a ton of comments (see the code!). I
didn't touch two functions (I was able to fix the bug without
going through them). I added comments to demonstrate that.

Either way, I fixed the edge artifact. Woo-hoo!