I implemented a new weighting system
(in fractal, the old one made no sense, sorry original author).

For the current setup, this amplified the angle vectors by a factor
of 24/7. Given this, an angle_scale of 2 should roughly give
the same level of wavy-ness as an old angle_scale of 7. This is
why I reduced the angle_scale.

I plan to make a bunch of new images using this new setup (and no
edge artifacts, woo-hoo!).