Split up into multiple files
I split up the Perlin_Flow_Field file into 5 smaller (and easier to
maintain/use) files, each of which is focused on a particular task in
the code.

The Main file houses the simulate, run, and load functions. These are
the functions that "drive" the code, so it makes sense for them to be
in main.

The Particles file houses the Particles class.

The Perlin file houses functions that generate the Perlin noise (fractal
, Perlin, etc.).

The Plot file contains the plotting functions (Plot_Flow, Plot_Vectors).

The Vector_Field file contains functions that set up the vector force
field from the Perlin noise. Namely rotate and Vector_Field.

I also renamed several functions to use capital letters, and swapped
np for numpy (idk why, I just wanted to do that).

the code itself has not changed... I just reorganized it. The old file
was over 700 lines long. It had become nearly impossible to maintain
and navigate. This segemented/multiple file approach should fix that.